+++ src/App.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { BrowserRouter, Link, Route, Routes } from "react-router-dom";
 import { Login } from "./pages/Login";
+import { Toppage } from "./pages/Toppage";
 
 function App() {
   return (
@@ -8,6 +9,7 @@ function App() {
       <BrowserRouter>
         <Link to="login">ログイン</Link>
         <Routes>
+          <Route path="/toppage" element={<Toppage />} />
           <Route path="/login" element={<Login />} />
         </Routes>
       </BrowserRouter>
